feat: add custom runtime exceptions

+++ vitalys-backend/src/main/java/com/vitalys/backend/controller/CargoController.java
@@ -33,12 +33,18 @@ public class CargoController {
 
     @DeleteMapping(path = "/cargos/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if(id == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         cargoRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
 
     @PutMapping(path = "/cargos/{id}")
     public ResponseEntity<CargoResponseDTO> update(@PathVariable Long id, @RequestBody CargoRequestDTO dto) {
+        if(id == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         return ResponseEntity.ok(cargoService.editar(id, dto));
     }
 }
